Switch TommyDS out for khash.
We want a hash table that can store more than 2^32 keys. TommyDS uses
"unsigned" a lot and it's not clear where that means size_t, where that
means uint_fast32_t, and where that means uint32_t. khash uses a typedef
of uint32_t by default, and should be far more easily convertible to use
size_t.

The only modifications done to khash.h thus far are adding AC_CATH and
AC_CAT, so that we can pass mushcoords as the name parameter to the kh_*
macros.