Move config.h.in to config/ to placate tup
tup doesn't (for now?) allow inputs in subdirectories to depend on
outputs in their parent directories. So if we want to, as we do, split
our source code into subdirectories, we can't have config.h created in
the project root directory: all subdirectories would depend on it.

As tup also doesn't (for now) allow creating output files in
subdirectories, we also have to move config.h.in to wherever we want
config.h created.

Since there's currently no other purpose for the directory, might as
well call it "config".