ICU: remove things we don't need
That is, stuff that isn't used anywhere, or conditional on #defines that
we know to always be false. A notable one I'm not currently sure about
is U_STATIC_IMPLEMENTATION: we might want to do something with that
depending on whether we're building a dynamic library or not.

utf{,8,16}.h are headers that we're going to be using, so they weren't
trimmed at all, at least yet.