space/load-string: fix textual whitespace handling
A test that'd break the old system would be trying to load " \nfoo" into
a non-allocated area.

This way should also be (a tiny bit) faster for "typical" inputs:
leading spaces are now handled at newlines, not on every space, so
spaces are (ditto) cheaper to load while newlines are (ditto) more
expensive.