irrelevize_subsumption_order: don't space higher
higher is a subsumee, it will disappear shortly anyway! There's no need
to care about what get_tight_bounds will do with it!

Plus, this could result in problems when looking at the consumee. (Or
when both are subsumees? Well, this argument suffices anyway.) Consider:

+-----++---+
| P +--| S |
|   |C +*--+
+---+----+-+

Where P is the one being placed, C is the consumee, S is a subsumee, and
* is a point contained both in C and S but whose correct data is in S.
Since S is above C, we'd copy the overlapping part from S to C and then
space it, including *, in S. As the consumee is taken care of first, P
would first get * from C, and then overwrite it with a space from S.
Oops: we replaced * with a space.