space/place-box: remove "subsumees" array
They can always be found from boxen by selecting those which are
contained in the consumer.

Other things this simplifies:

- No more subsume_contains. Although we need to do one similar pass at
  the end, to grab the best consumee and in case there are no other
  subsumees.

- No more null check in min_max_size.

- No more removed_offsets complexity in consume_and_subsume.

- No more mallocs in the whole file!

- Bottom-up subsuming in consume_and_subsume "for free", which also
  allows us to exclude some copies in what was
  irrelevize_subsumption_order.

The removal of subsume_contains may also have a slight "correctness"
impact: it increased used_cells for all the contained boxes, thus
causing them to be counted twice. This could result in some "wrong"
decisions – correctness and wrong in quotes because things would work
fine, but the heuristics weren't making the intended decisions.