Handle rare errors by callback: abort() by default
The callback is stored in mushspace to avoid having to pass it around
all the time. mushspace_signal calls it and then, in case it returns,
calls the default callback, which prints a message to stderr and aborts.

Basically the callback should longjmp, throw an exception, or exit the
program. If it doesn't, we avoid undefined behaviour by exiting the
program ourselves instead.

Major consequences:

- mushcursor_init can now work like mushspace_init, since it doesn't
  have the option of returning an error code. For the ability to place a
  cursor in an empty space, it now accepts a zero delta as a special
  case.

- No more MUSHERR_NONE, yay.

- mushspace_place_box_for is now mushspace_place_box_put and also
  handles writing into the placed AABB. This means that mushspace_put
  can just tail call it without USE_BAKAABB, and it can be the one that
  signals the MUSHERR_INVALIDATION_FAILURE, so mushspace_put doesn't
  need to know about it at all.

Note that Clang doesn't yet understand _Noreturn, so we #define it away:
http://llvm.org/bugs/show_bug.cgi?id=13818

Oh, and this is something like a 5-6% performance win on average.